rename(guest): security_policy.rs -> resolved_security.rs
Keeps the word this module exists to signal: readonly_paths/sys_mount_options
here are already-resolved literal values, not rules waiting to be applied —
dropping "resolved" left "security_policy" reading like the latter, and
close enough to the host's unrelated SecurityOptions (shim sandboxing) to
invite confusion across the two crates. Pure rename: git mv + 4 import sites,
no logic changes.
