+++ src/guest/src/container/lifecycle.rs
@@ -4,7 +4,7 @@
 //! Follows the OCI Runtime Specification.
 
 use super::command::ContainerCommand;
-use super::security_policy::ResolvedSecurityPolicy;
+use super::resolved_security::ResolvedSecurityPolicy;
 use super::spec::{ContainerDevices, UserMount};
 use super::stdio::{ContainerStdio, InitIo};
 use super::{console_socket, kill, spec, start};
